requirements_dev.txt
# commands
invoke==2.2.0

# pre-commit
pre-commit==3.5.0

# testing
pytest==7.4.3
coverage==7.6.9

# linting
ruff==0.1.3

# docs
mkdocs==1.5.3
mkdocs-material==9.4.6
mkdocstrings-python==1.12.2

#logging
loguru~=0.7.3
wandb~=0.19.2

#cli
typer==0.15.1

#env
python-dotenv==1.0.1

#config
omegaconf~=2.3.0
hydra-core~=1.3.2

#ml
scikit_learn==1.6.0
torch~=2.5.1
torchvision==0.20.1
matplotlib~=3.10.0
scikit-learn~=1.6.0
